Made Leaderboard page fully SSR
* refactor: used a wrapper to wrap leaderboard

* chore: ran pnpm format

* refactor: made the /leaderboard route fully SSR except searchbar

* fix: forced the LeaderboardWrapper to be dynamic

* fix: fixed build by replacing use of useSearchParams by searchParams provided by page

* fix: fixed type error for role in Searchbar for searchParams

* fix: fixed unintensional css changes in top leaderboard
