+++ .prettierrc.json
@@ -5,10 +5,6 @@
   "semi": true,
   "jsxSingleQuote": false,
   "arrowParens": "always",
-  "tailwindFunctions": [
-    "classNames"
-  ],
-  "plugins": [
-    "prettier-plugin-tailwindcss"
-  ]
+  "tailwindFunctions": ["classNames"],
+  "plugins": ["prettier-plugin-tailwindcss"]
 }
