+++ app/globals.css
@@ -23,7 +23,7 @@ body {
 }
 
 .ContentSection {
-  @apply max-w-screen-sm mx-auto;
+  @apply mx-auto max-w-screen-sm;
   @apply py-8;
   @apply px-6;
 }
@@ -38,7 +38,7 @@ body {
 
 @screen md {
   .ContentSection {
-    @apply max-w-screen-lg mx-auto;
+    @apply mx-auto max-w-screen-lg;
     @apply py-32;
     @apply w-2/3;
   }
@@ -91,7 +91,7 @@ html.dark {
 .tooltip .tooltip-text {
   visibility: hidden;
   opacity: 0;
-  @apply bg-black/75 backdrop-blur text-white text-center p-2 rounded absolute z-50 text-sm block transition-opacity whitespace-nowrap pointer-events-none;
+  @apply pointer-events-none absolute z-50 block whitespace-nowrap rounded bg-black/75 p-2 text-center text-sm text-white backdrop-blur transition-opacity;
 }
 
 .tooltip .tooltip-left {
