--- app/leaderboard/loading.tsx
@@ -1,5 +1,0 @@
-import LoadingText from "@/components/LoadingText";
-
-export default function Loading() {
-  return <LoadingText text="Ranking the contributors" />;
-}
